requirements.txt
# Core dependencies
PyQt6>=6.5.0,<7.0.0
requests>=2.28.0,<3.0.0
beautifulsoup4>=4.11.0,<5.0.0
PyPDF2>=3.0.0,<4.0.0
python-docx>=0.8.11,<1.0.0
nltk>=3.8.0,<4.0.0
python-dotenv>=1.0.0,<2.0.0
psutil>=5.9.0,<6.0.0
rich>=13.0.0,<14.0.0

# ML dependencies
wandb>=0.15.0,<1.0.0
torch>=2.0.0,<3.0.0
transformers>=4.30.0,<5.0.0
peft>=0.4.0,<1.0.0
datasets>=2.12.0,<3.0.0
accelerate>=0.20.0,<1.0.0
bitsandbytes>=0.39.0,<1.0.0
tqdm>=4.65.0,<5.0.0

# Vector DB and embedding dependencies
faiss-cpu>=1.7.4,<2.0.0
sentence-transformers>=2.2.2,<3.0.0
chromadb>=0.4.13,<1.0.0
# Support for more complex data structures
networkx>=3.1,<4.0.0
pandas>=2.0.0,<3.0.0
# Industry-specific packages
scikit-learn>=1.2.0,<2.0.0
matplotlib>=3.7.0,<4.0.0
